--- apps/kg/models/pytorch/__init__.py
@@ -1,1 +1,0 @@
-from .tensor_models import PBGKEModel
